Extract Plotting Boilerplate to Separate File
