--- test/mocks/inference/pyseir_default_parameters.py
@@ -1,11 +1,0 @@
-pyseir_default_parameters = {
-    # Produces serial_time = 5
-    "sigma": 0.333,
-    "delta": 0.25,
-    "test_fraction": 0.1,
-    # Used when inferring new inferring new hospital admissions
-    "hospitalization_length_of_stay_general": 7.04,
-    "hospitalization_length_of_stay_icu": 8.00,
-    "hospitalization_rate_general": 0.0201,
-    "hospitalization_rate_icu": 0.00606,
-}
