--- test/pyseir/infer_rt_test.py
@@ -1,12 +1,0 @@
-import pandas as pd
-import structlog
-from pyseir.rt import infer_rt
-
-
-def test_replace_outliers_on_last_day():
-    x = pd.Series([10, 10, 10, 500], [0, 1, 2, 3])
-
-    results = infer_rt.replace_outliers(x, structlog.getLogger(), local_lookback_window=3)
-
-    expected = pd.Series([10, 10, 10, 10], [0, 1, 2, 3])
-    pd.testing.assert_series_equal(results, expected)
